Fix CRON (#24)
* Use docker-compose to run database setup
* Build sidekiq image in test script
* Fix db migration startup script
* Let Runit deal with services

Sidekiq was started by a blocking script in my_init. This lead to the init process not getting to run Runit, which is in charge of CRON among other things. To fix it, I moved Sidekiq to be run by Runit as well. I also moved the postfix start script, while I was at it.