Fix CRON environment (#28)
CRON seems to only be able to access the environment variables if given access to `/etc/container_environment*`. This is necessary for the rake tasks run by CRON to be able to do their jobs. Herewith, we shall grant access by adding the  `app` user to the `docker_env` group. ✍️